Moved AmigaOS4 specific files to platform/amigaos
