+++ gemrb/core/CMakeLists.txt
@@ -5,7 +5,7 @@ IF (ANDROID)
 ELSEIF (WIN32)
 	set (PLATFORM_SRC System/Logger/Win32Console.cpp)
 ELSEIF (AmigaOS4)
-        set (PLATFORM_SRC System/OS4Helpers.cpp)
+        set (PLATFORM_SRC ../../platforms/amigaos/OS4Helpers.cpp)
 ELSE ()
 	set (PLATFORM_SRC)
 ENDIF ()
