+++ platforms/amigaos/OS4Helpers.cpp
