+++ platforms/amigaos/OS4Helpers.h
